+++ common/src/main/java/com/hvost/blog/support/PostService.java
@@ -58,7 +58,7 @@ public class PostService {
         return null;
     }
 
-  public Page<Post> getPublishPostByCategory(Integer category, Pageable pageRequest){
+  public Page<Post> getPublishPostByCategory(Long category, Pageable pageRequest){
     return postRepository.findByCategoryPost_Id(category, pageRequest);
   }
